+++ qqrobot-spring-boot-starter/src/main/java/de/honoka/qqrobot/starter/common/RobotBeanHolder.java
@@ -2,7 +2,7 @@ package de.honoka.qqrobot.starter.common;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
-import de.honoka.qqrobot.framework.Framework;
+import lombok.Getter;
 import org.springframework.context.ApplicationContext;
 import org.springframework.stereotype.Component;
 
@@ -11,6 +11,7 @@ import javax.annotation.Resource;
 /**
  * 解决循环依赖
  */
+@Getter
 @Component
 public class RobotBeanHolder {
 
@@ -21,15 +22,4 @@ public class RobotBeanHolder {
             .serializeNulls()
             .setDateFormat("yyyy-MM-dd HH:mm:ss")
             .create();
-
-    private Framework<?> framework;
-
-    public Framework<?> getFramework() {
-        if(framework != null) return framework;
-        synchronized(this) {
-            if(framework != null) return framework;
-            framework = applicationContext.getBean(Framework.class);
-            return framework;
-        }
-    }
 }