+++ qqrobot-spring-boot-starter/src/main/java/de/honoka/qqrobot/starter/component/controller/BasicRobotController.java
@@ -3,8 +3,9 @@ package de.honoka.qqrobot.starter.component.controller;
 import de.honoka.qqrobot.framework.model.RobotMessage;
 import de.honoka.qqrobot.starter.common.annotation.Command;
 import de.honoka.qqrobot.starter.common.annotation.RobotController;
-import de.honoka.qqrobot.starter.component.RobotAttributes;
+import de.honoka.qqrobot.starter.component.MessageExecutor;
 import de.honoka.sdk.util.various.ImageUtils;
+import org.springframework.context.annotation.Lazy;
 
 import javax.annotation.Resource;
 import java.io.InputStream;
@@ -13,12 +14,13 @@ import java.io.InputStream;
 @RobotController
 public class BasicRobotController {
 
+    @Lazy
     @Resource
-    private RobotAttributes robotAttributes;
+    private MessageExecutor messageExecutor;
 
     @Command("菜单")
     public RobotMessage<InputStream> menu() {
         return RobotMessage.image(ImageUtils.textToImageByLength(
-                robotAttributes.menu, 30));
+                messageExecutor.getMenu(), 30));
     }
 }