+++ qqrobot-spring-boot-starter/src/main/java/de/honoka/qqrobot/starter/framework/tester/config/TesterProperties.java
@@ -16,4 +16,6 @@ public class TesterProperties {
             "tester-framework", "image").toString();
 
     private String webPrefix = "/tester-framework";
+
+    private Long groupNumber = 10000L;
 }